requirements.txt
# Core web framework
fastapi
uvicorn
python-dotenv

# Authentication
passlib
fastapi-jwt
authlib

# Database
pymilvus

# Data validation
pydantic

# LangChain and AI integrations
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
transformers
huggingface-hub
faiss-cpu

# PDF and YouTube extraction
PyPDF2
pytubefix

## code quality
pip install pre-commit==4.2.0
